+++ XAMLTest/Host/VisualTreeService.cs
@@ -28,7 +28,7 @@ using Window = System.Windows.Window;
 
 namespace XamlTest.Host
 {
-    internal class VisualTreeService : Protocol.ProtocolBase
+    internal partial class VisualTreeService : Protocol.ProtocolBase
     {
         private static Guid Initialized { get; } = Guid.NewGuid();
 
@@ -587,157 +587,6 @@ namespace XamlTest.Host
             return reply;
         }
 
-        public override async Task<KeyboardFocusResult> MoveKeyboardFocus(KeyboardFocusRequest request, ServerCallContext context)
-        {
-            KeyboardFocusResult reply = new();
-            await Application.Dispatcher.InvokeAsync(() =>
-            {
-                if (!(GetCachedElement<DependencyObject>(request.ElementId) is IInputElement element))
-                {
-                    reply.ErrorMessages.Add("Could not find element");
-                    return;
-                }
-                if (element is DependencyObject @do &&
-                    Window.GetWindow(@do) is Window window)
-                {
-                    if (!ActivateWindow(window))
-                    {
-                        reply.ErrorMessages.Add($"Failed to activate window.");
-                        return;
-                    }
-                }
-                else
-                {
-                    reply.ErrorMessages.Add($"Failed to find parent window.");
-                }
-
-                if (Keyboard.Focus(element) != element)
-                {
-                    reply.ErrorMessages.Add($"Failed to move focus to element {element}");
-                }
-                if (element is UIElement uIElement)
-                {
-                    uIElement.Focus();
-                }
-            });
-            return reply;
-        }
-
-        public override async Task<InputResponse> SendInput(InputRequest request, ServerCallContext context)
-        {
-            InputResponse reply = new();
-            await Application.Dispatcher.Invoke(async () =>
-            {
-                try
-                {
-                    IntPtr windowHandle = IntPtr.Zero;
-                    if (!(GetCachedElement<DependencyObject>(request.ElementId) is { } element))
-                    {
-                        reply.ErrorMessages.Add("Could not find element");
-                        return;
-                    }
-
-                    Window window = Window.GetWindow(element);
-                    if (window is null)
-                    {
-                        reply.ErrorMessages.Add("Failed to find parent window");
-                        return;
-                    }
-                    windowHandle = new WindowInteropHelper(window).EnsureHandle();
-
-                    if (!ActivateWindow(window))
-                    {
-                        reply.ErrorMessages.Add($"Failed to active window");
-                        return;
-                    }
-
-                    if (windowHandle != IntPtr.Zero)
-                    {
-                        foreach (MouseData mouseData in request.MouseData)
-                        {
-                            switch (mouseData.Event)
-                            {
-                                case MouseData.Types.MouseEvent.MoveToElement:
-                                    if (element is FrameworkElement frameworkElement)
-                                    {
-                                        Rect coordinates = GetCoordinates(frameworkElement);
-                                        Position position = Position.Center;
-                                        if (!string.IsNullOrEmpty(mouseData.Value))
-                                        {
-                                            Enum.TryParse(mouseData.Value, out position);
-                                        }
-                                        Point location = position switch
-                                        {
-                                            Position.TopLeft => coordinates.TopLeft,
-                                            Position.TopRight => coordinates.TopRight,
-                                            Position.BottomRight => coordinates.BottomRight,
-                                            Position.BottomLeft => coordinates.BottomLeft,
-                                            _ => coordinates.Center()
-                                        };
-                                        Input.MouseInput.MoveCursor(location);
-                                    }
-                                    break;
-                                case MouseData.Types.MouseEvent.LeftDown:
-                                    Input.MouseInput.LeftDown();
-                                    break;
-                                case MouseData.Types.MouseEvent.LeftUp:
-                                    Input.MouseInput.LeftUp();
-                                    break;
-                                case MouseData.Types.MouseEvent.MiddleDown:
-                                    Input.MouseInput.MiddleDown();
-                                    break;
-                                case MouseData.Types.MouseEvent.MiddleUp:
-                                    Input.MouseInput.MiddleUp();
-                                    break;
-                                case MouseData.Types.MouseEvent.RightDown:
-                                    Input.MouseInput.RightDown();
-                                    break;
-                                case MouseData.Types.MouseEvent.RightUp:
-                                    Input.MouseInput.RightUp();
-                                    break;
-                                case MouseData.Types.MouseEvent.Delay:
-                                    if (!string.IsNullOrEmpty(mouseData.Value) &&
-                                        int.TryParse(mouseData.Value, out int millisecondsDelay))
-                                    {
-                                        await Task.Delay(TimeSpan.FromMilliseconds(millisecondsDelay));
-                                    }
-                                    break;
-                            }
-                        }
-                        if (request.KeyboardData.Any() && element is IInputElement inputElement)
-                        {
-                            if (Keyboard.Focus(inputElement) != element)
-                            {
-                                reply.ErrorMessages.Add($"Failed to move focus to element {element}");
-                                return;
-                            }
-                        }
-                        await Task.Run(async () =>
-                        {
-                            foreach (KeyboardData keyboardData in request.KeyboardData)
-                            {
-                                if (!string.IsNullOrEmpty(keyboardData.TextInput))
-                                {
-                                    Input.KeyboardInput.SendKeysForText(windowHandle, keyboardData.TextInput);
-                                }
-                                if (keyboardData.Keys.Any())
-                                {
-                                    Input.KeyboardInput.SendKeys(windowHandle, keyboardData.Keys.Cast<Key>().ToArray());
-                                }
-                                await Task.Delay(10);
-                            }
-                        });
-                    }
-                }
-                catch (Exception e)
-                {
-                    reply.ErrorMessages.Add(e.ToString());
-                }
-            });
-
-            return reply;
-        }
-
         public override Task<ShutdownResponse> Shutdown(ShutdownRequest request, ServerCallContext context)
         {
             ShutdownResponse reply = new();
@@ -780,74 +629,6 @@ namespace XamlTest.Host
                 reply.ErrorMessages.Add(e.ToString());
             }
             return Task.FromResult(reply);
-        }
-
-        public override async Task<EventRegistrationResponse> RegisterForEvent(EventRegistrationRequest request, ServerCallContext context)
-        {
-            EventRegistrationResponse reply = new()
-            {
-                EventId = Guid.NewGuid().ToString()
-            };
-            await Application.Dispatcher.InvokeAsync(() =>
-            {
-                DependencyObject? element = GetCachedElement<DependencyObject>(request.ElementId);
-                if (element is null)
-                {
-                    reply.ErrorMessages.Add("Could not find element");
-                    return;
-                }
-
-                if (element.GetType().GetEvent(request.EventName) is { } eventInfo)
-                {
-                    EventRegistrar.Regsiter(reply.EventId, eventInfo, element);
-                }
-            });
-            return reply;
-        }
-
-        public override Task<EventUnregisterResponse> UnregisterForEvent(EventUnregisterRequest request, ServerCallContext context)
-        {
-            EventUnregisterResponse reply = new();
-            if (!EventRegistrar.Unregister(request.EventId))
-            {
-                reply.ErrorMessages.Add("Failed to unregister event");
-            }
-            return Task.FromResult(reply);
-        }
-
-        public override Task<EventInvocationsResponse> GetEventInvocations(EventInvocationsQuery request, ServerCallContext context)
-        {
-            EventInvocationsResponse reply = new()
-            {
-                EventId = request.EventId,
-            };
-            var invocations = EventRegistrar.GetInvocations(request.EventId);
-            if (invocations is null)
-            {
-                reply.ErrorMessages.Add("Event was not registered");
-            }
-            else
-            {
-                reply.EventInvocations.AddRange(
-                    invocations.Select(
-                        array =>
-                        {
-                            EventInvocation rv = new();
-                            rv.Parameters.AddRange(array.Select(item => GetItemString(item)));
-                            return rv;
-                        }));
-            }
-
-
-            return Task.FromResult(reply);
-
-            string GetItemString(object item)
-            {
-                return Serializer.Serialize(item.GetType(), item)
-                    ?? item?.ToString()
-                    ?? item?.GetType().FullName
-                    ?? "<null>";
-            }
         }
 
         private Assembly? CurrentDomain_AssemblyResolve(object? sender, ResolveEventArgs args)
